requirements.txt
openai
python-dotenv
ipython
ipywidgets
gradio
requests
pandas
folium
streamlit
gradio_client
beautifulsoup4